requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==3.4.8
cycler==0.11.0
eight==1.0.1
fonttools==4.36.0
frozenlist==1.3.1
future==0.18.2
idna==3.3
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
multidict==6.0.2
mypy==0.971
mypy-extensions==0.4.3
numpy==1.23.2
openleadr==0.5.26
packaging==21.3
Pillow==9.2.0
pycparser==2.21
pyOpenSSL==19.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.1
signxml-openadr==2.9.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
xmltodict==0.13.0
yarl==1.8.1
